+++ libc-test/src/functionalext/supplement/unistd/unlinkat.c
@@ -27,6 +27,10 @@ void unlinkat_0100(void)
 {
     char path[128] = {0};
     char *cwd = getcwd(path, sizeof(path));
+    if (!cwd) {
+        t_error("%s getcwd failed\n", __func__);
+        return;
+    }
     strcat(path, "/file.txt");
 
     int fd = open(path, O_CREAT, 0664);
@@ -49,6 +53,10 @@ void unlinkat_0200(void)
     errno = 0;
     char path[128] = {0};
     char *cwd = getcwd(path, sizeof(path));
+    if (!cwd) {
+        t_error("%s getcwd failed\n", __func__);
+        return;
+    }
     strcat(path, "/file.txt");
     int result = unlinkat(AT_FDCWD, path, AT_REMOVEDIR);
     if (result == 0) {
